refactor: 내 위시 페이지 불필요한 client wrapper 삭제
